backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.32.0
sqlalchemy==2.0.36
psycopg2-binary==2.9.10
pydantic==2.8.2
python-jose==3.3.0
passlib[bcrypt]==1.7.4
alembic==1.13.2
pydantic-settings==2.6.1
email-validator==2.3.0

